pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "svlm"
version = "0.1.0"
description = "Post-hoc calibration framework for small vision-language models."
readme = "README.md"
authors = [{name = "SVLM Team"}]
requires-python = ">=3.9"
dependencies = [
    "hydra-core>=1.3",
    "omegaconf>=2.3",
    "torch>=2.1",
    "tqdm>=4.65",
    "numpy>=1.24",
    "pydantic>=2.4"
]

[project.optional-dependencies]
dev = ["pytest>=7.4", "ruff>=0.1"]
